basket/basket-service/src/test/kotlin/jp/inaba/basket/service/application/basket/CreateBasketInteractorTest.kt
package jp.inaba.basket.service.application.basket

import com.github.michaelbull.result.Err
import com.github.michaelbull.result.Ok
import jp.inaba.basket.api.domain.basket.BasketId
import jp.inaba.basket.api.domain.basket.CreateBasketCommand
import jp.inaba.basket.api.domain.basket.CreateBasketError
import jp.inaba.basket.service.application.command.basket.CreateBasketInteractor
import jp.inaba.basket.service.domain.basket.CanCreateBasketVerifier
import jp.inaba.basket.service.domain.basket.InternalCreateBasketCommand
import jp.inaba.identity.api.domain.user.UserId
import org.axonframework.commandhandling.gateway.CommandGateway
import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.Test
import org.mockito.InjectMocks
import org.mockito.Mock
import org.mockito.Mockito
import org.mockito.MockitoAnnotations

class CreateBasketInteractorTest {
    @Mock
    private lateinit var canCreateBasketVerifier: CanCreateBasketVerifier

    @Mock
    private lateinit var commandGateway: CommandGateway

    @InjectMocks
    private lateinit var sut: CreateBasketInteractor

    @BeforeEach
    fun before() {
        MockitoAnnotations.openMocks(this)
    }

    @Test
    fun `ユーザーが存在_買い物かごを作成_InternalCommandが配送`() {
        val userId = UserId()
        val command = CreateBasketCommand(userId)
        Mockito.`when`(canCreateBasketVerifier.checkUserExits(userId))
            .thenReturn(Ok(Unit))

        val result = sut.handle(command)

        assert(result.isOk())
        val expectCommand = InternalCreateBasketCommand(BasketId(userId))
        Mockito.verify(commandGateway, Mockito.only()).sendAndWait<Any>(expectCommand)
    }

    @Test
    fun `ユーザーが存在しない_買い物かごを作成_InternalCommandが配送されずエラーが返る`() {
        val userId = UserId()
        val command = CreateBasketCommand(userId)
        Mockito.`when`(canCreateBasketVerifier.checkUserExits(userId))
            .thenReturn(Err(CreateBasketError.USER_NOT_FOUND))

        val result = sut.handle(command)

        assert(!result.isOk())
        assert(result.errorCode == CreateBasketError.USER_NOT_FOUND.errorCode)
        val expectCommand = InternalCreateBasketCommand(BasketId(userId))
        Mockito.verify(commandGateway, Mockito.never()).sendAndWait<Any>(expectCommand)
    }
}
